Code: Optimized

+++ src/main/java/com/kakaouo/mods/yuunalive/entities/ai/goal/YuunaLivePlayerFollowOwnerGoal.java
@@ -6,7 +6,9 @@ import net.minecraft.block.BlockState;
 import net.minecraft.block.LeavesBlock;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.ai.goal.Goal;
-import net.minecraft.entity.ai.pathing.*;
+import net.minecraft.entity.ai.pathing.EntityNavigation;
+import net.minecraft.entity.ai.pathing.LandPathNodeMaker;
+import net.minecraft.entity.ai.pathing.PathNodeType;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.WorldView;
 
@@ -110,7 +112,7 @@ public class YuunaLivePlayerFollowOwnerGoal extends Goal {
         } else if (!this.canTeleportTo(new BlockPos(x, y, z))) {
             return false;
         } else {
-            this.entity.refreshPositionAndAngles((double)x + 0.5D, (double)y, (double)z + 0.5D, this.entity.getYaw(), this.entity.getPitch());
+            this.entity.refreshPositionAndAngles((double)x + 0.5D, y, (double)z + 0.5D, this.entity.getYaw(), this.entity.getPitch());
             this.navigation.stop();
             return true;
         }
